# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Payment {
  matricNumber: string;
  studentName: string;
  session: string;
  amount: number;
  code: string;
  timestamp: string;
}

export interface PaymentInput {
  /** @minLength 1 */
  matricNumber: string;
  /** @minLength 1 */
  studentName: string;
  /** @minLength 1 */
  session: string;
  /** @minimum 0 */
  amount: number;
}

export interface StandingSession {
  session: string;
  amount: number;
  code: string;
  timestamp: string;
}

export interface Standing {
  matricNumber: string;
  studentName: string;
  paidSessions: StandingSession[];
  outstandingSessions: string[];
  totalPaid: number;
  totalOwed: number;
}

export interface ErrorResponse {
  error: string;
}