# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  matricNumber: string;
  studentName: string;
  session: string;
  amount: number;
  code: string;
  timestamp: Date;
}