# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  /** @minLength 1 */
  matricNumber: string;
  /** @minLength 1 */
  studentName: string;
  /** @minLength 1 */
  session: string;
  /** @minimum 0 */
  amount: number;
}