# lib/api-zod/src/generated/types/standing.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StandingSession } from './standingSession';

export interface Standing {
  matricNumber: string;
  studentName: string;
  paidSessions: StandingSession[];
  outstandingSessions: string[];
  totalPaid: number;
  totalOwed: number;
}