# lib/api-zod/src/generated/types/standingSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StandingSession {
  session: string;
  amount: number;
  code: string;
  timestamp: Date;
}